Fixed Model increment, dates. Fixed Migrations,. Added enum Language, support request lang. Working added userToken method

+++ backend/database/migrations/2018_02_16_160336_create_tags_table.php
@@ -15,7 +15,7 @@ class CreateTagsTable extends Migration
     {
         Schema::create('tags', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('group_id');
+            $table->unsignedBigInteger('group_id');
             $table->string('name')->unique();
             $table->string('hash');
             $table->softDeletes();
